refactor(system): move archive password request to external models
relocate ArchivePasswordRequest from services/io to models/external and update namespace imports in core and ui
