+++ AvatarExplorer.Core/Services/System/AvatarExplorerApp.cs
@@ -1,5 +1,6 @@
 using System.Collections.Immutable;
 using AvatarExplorer.Core.Data.Paths;
+using AvatarExplorer.Core.Models.External;
 using AvatarExplorer.Core.Models.Items;
 using AvatarExplorer.Core.Models.System;
 using AvatarExplorer.Core.Services.Database;
