+++ AvatarExplorer.UI/Views/MainWindow.axaml.cs
@@ -15,6 +15,7 @@ using AvatarExplorer.Core.Data.Links;
 using AvatarExplorer.Core.Data.Paths;
 using AvatarExplorer.Core.Extensions;
 using AvatarExplorer.Core.Localization;
+using AvatarExplorer.Core.Models.External;
 using AvatarExplorer.Core.Models.Items;
 using AvatarExplorer.Core.Models.System;
 using AvatarExplorer.Core.Services.IO;
